fix: Resolve comprehensive CI/CD pipeline failures and code quality issues
This commit addresses multiple critical issues that were causing CI/CD pipeline
failures and improves overall code quality and test reliability.

**Core CI/CD Fixes:**
- Fix undefined o2types.SubscriptionSpec by creating missing o2-client/go.mod
- Remove unused imports and variables causing lint failures
- Fix context leaks in iperf.go with proper cleanup handling
- Resolve assignment mismatches in vxlan.go for SecureExecute calls
- Add missing security package imports across test files

**Placement Algorithm Improvements:**
- Fix scoring calculations for latency and throughput metrics
- Add penalty system for impossible placement scenarios
- Implement proper sorting of placement alternatives by score
- Update test data to align with thesis performance targets
- All placement tests now pass with correct QoS requirements

**Enhanced Security Scanning:**
- Improve gosec SARIF file generation in CI workflow
- Add automatic jq installation for SARIF processing
- Better error handling and validation for security scans
- Fixed file path handling and debugging output

**Test Framework Enhancements:**
- Fix float to int conversion issues in integration tests
- Resolve duplicate import declarations
- Add proper security imports to test utilities
- Fix startTime variable scope issues in performance tests

**Build System:**
- Create o2-client module with proper go.mod
- Add replace directive for local module resolution
- Resolve all go vet and build compilation errors
- Ensure all modules build successfully

**Verification:**
- All placement tests pass ✅
- Core builds compile without errors ✅
- Go vet checks pass ✅
- Module dependencies properly resolved ✅

This resolves CI failures related to code quality, unit tests, build processes,
and security scanning while maintaining thesis performance targets.

+++ tests/integration/nephio_integration_test.go
@@ -553,7 +553,7 @@ var _ = Describe("Nephio Package Generation and Deployment Tests", func() {
 
 			totalGenerationTime := time.Since(generationStart)
 
-			Expect(successCount).To(BeNumerically(">=", packageCount*0.95),
+			Expect(successCount).To(BeNumerically(">=", int(float64(packageCount)*0.95)),
 				"At least 95% of packages should generate successfully")
 
 			Expect(totalGenerationTime).To(BeNumerically("<=", 5*time.Minute),
